refactor(review-execution): finish reviewContextAction Strangler
The Strangler Fig in reviewContext/reviewContextAction.{schema,guard}.ts
was carrying a 'will be removed once all imports are updated' comment.
Updates the imports — all 6 consumers now import directly from the
canonical locations.

- New files for the result-specific part (kept separately because it is
  unique to reviewContext, not a re-export):
  - entities/reviewContext/reviewContextResult.schema.ts
  - entities/reviewContext/reviewContextResult.guard.ts

- Consumers migrated:
  - reviewContext.gateway.ts, reviewContext.schema.ts, reviewContext.ts
  - interface-adapters/gateways/reviewContext.fileSystem.gateway.ts
  - usecases/mcp/addAction.usecase.ts
  - tests/stubs/reviewContextGateway.stub.ts
  Each splits the type import: ReviewContextAction from reviewAction.ts,
  ReviewContextResult from the new reviewContextResult.schema.ts.

- Strangler files deleted:
  - reviewContextAction.schema.ts (just re-exported reviewAction)
  - reviewContextAction.guard.ts (just re-exported guards)
  - Their tests (action portion was redundant with reviewAction.guard.test.ts;
    result portion ported to new reviewContextResult.{schema,guard}.test.ts).

The 'ReviewContextAction = ReviewAction' alias in reviewAction.ts is
kept as a soft-deprecation — eliminating that name entirely is a
separate rename (out of scope for HS-11).

Resolves HS-11 from the DDD hot spots.
