+++ src/modules/review-execution/entities/reviewContext/reviewContext.ts
@@ -1,4 +1,5 @@
-import type { ReviewContextAction, ReviewContextResult } from './reviewContextAction.schema.js'
+import type { ReviewContextAction } from '@/modules/review-execution/entities/reviewAction/reviewAction.js'
+import type { ReviewContextResult } from './reviewContextResult.schema.js'
 
 export interface DiffMetadata {
   baseSha: string
