+++ src/modules/review-execution/interface-adapters/gateways/reviewContext.fileSystem.gateway.ts
@@ -9,7 +9,8 @@ import type {
   ReviewContextProgress,
 } from '@/modules/review-execution/entities/reviewContext/reviewContext.js'
 import { buildAgentInstructions } from '@/services/agentInstructionsBuilder.js'
-import type { ReviewContextAction, ReviewContextResult } from '@/modules/review-execution/entities/reviewContext/reviewContextAction.schema.js'
+import type { ReviewContextAction } from '@/modules/review-execution/entities/reviewAction/reviewAction.js'
+import type { ReviewContextResult } from '@/modules/review-execution/entities/reviewContext/reviewContextResult.schema.js'
 
 export class ReviewContextFileSystemGateway implements ReviewContextGateway {
   getFilePath(localPath: string, mergeRequestId: string): string {
